DX: simplify plugin's code
- Extract the `DiscourseLoginClientStrategy` into its own file
- Remove default methods from the `DiscourseLoginClientAuthenticator`
